haystack/app/requirements.txt
beautifulsoup4==4.6.0
# Flask==0.12
Flask-Cors==3.0.2
microdata==0.6.1
PyMySQL==0.7.11
pyquery==1.2.17
requests==2.18.4
SQLAlchemy==1.1.5
numpy==1.13.0
pandas==0.21.0
Pillow==4.0.0
scipy==1.0.0
scikit-learn==0.18.2
selenium==3.5.0